infr/requirements.txt
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.7.0
fastapi>=0.110.0
uvicorn>=0.30.0
